+++ CHANGELOG.md
@@ -12,9 +12,15 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Added
 
+- Added `replace_item` to `DesignerView`, `Section`, `Container`, `ActionRow`, &
+  `MediaGallery` ([#3032](https://github.com/Pycord-Development/pycord/pull/3032))
+
 ### Changed
 
 ### Fixed
+
+- Fixed core issues with modifying items in `Container` and `Section`
+  ([#3032](https://github.com/Pycord-Development/pycord/pull/3032))
 
 ### Deprecated